backend/pyproject.toml
[project]
name = "numeros-backend"
version = "1.0.0"
description = "Numeros Love/Compatibility API - Numerology and Astrology matching"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Django Core
    "django>=5.2",
    "djangorestframework>=3.15",
    "django-cors-headers>=4.3",
    "django-filter>=24.0",

    # Authentication
    "djangorestframework-simplejwt>=5.3",

    # Database
    "psycopg[binary]>=3.1",

    # Astrology
    "pyswisseph>=2.10",
    "pytz>=2024.1",
    "timezonefinder>=6.5",

    # Utilities
    "python-dotenv>=1.0",
    "gunicorn>=21.0",

    # File uploads
    "pillow>=10.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-django>=4.8",
    "pytest-cov>=4.1",
    "black>=24.0",
    "ruff>=0.2",
]

prod = [
    # WebSocket (for messaging)
    "channels>=4.0",
    "channels-redis>=4.2",
    # S3 uploads
    "boto3>=1.34",
    "django-storages>=1.14",
]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.development"
python_files = ["test_*.py", "*_test.py"]
